+++ R/witch_scen.R
@@ -3,7 +3,6 @@
 #' Returns a scenario name
 #'
 #' @family WITCH helper functions
-#' @seealso \code{\link{witch_query}}}.
 #'
 #' @export
 witch_scen_name <- function(gdxfilename) {